[python] create plot files, fix SettingWithCopyWarning
