+++ python/dalex/dalex/instance_level/_ceteris_paribus/plot.py
@@ -1,0 +1,18 @@
+def tooltip_text(obs, r=None):
+    temp = ""
+    if r is not None:
+        for var in obs.index:
+            if var == "_yhat_":
+                temp += "prediction:<br>" + "- before: " + str(obs[var]) + "<br>" + "- after: " +\
+                        str(r[var]) + "<br><br>"
+            elif var == r['_vname_']:
+                temp += var + ": " + str(r['_xhat_']) + "</br>"
+            else:
+                temp += var + ": " + str(obs[var]) + "</br>"
+    else:
+        for var in obs.index:
+            if var == "_yhat_":
+                temp += "prediction:" + str(obs[var]) + "<br><br>"
+            else:
+                temp += var + ": " + str(obs[var]) + "</br>"
+    return temp
