+++ python/dalex/dalex/instance_level/_shap/object.py
@@ -1,6 +1,7 @@
 import numpy as np
 from plotly.subplots import make_subplots
 
+from dalex.instance_level._shap.plot import prepare_data_for_shap_plot
 from .checks import *
 from .utils import shap
 from ..._explainer.theme import get_break_down_colors
@@ -59,8 +60,6 @@ class Shap:
         :param title: str, the plot's title
         """
 
-        deleted_indexes = []
-
         # are there any other objects to plot?
         if objects is None:
             n = 1
@@ -86,6 +85,7 @@ class Shap:
                 _prediction_list += [ob.prediction]
 
         # TODO: add intercept and prediction list update for multi-class
+        # deleted_indexes = []
         # for i in range(n):
         #     result = _result_list[i]
         #
@@ -178,49 +178,3 @@ class Shap:
             'modeBarButtonsToRemove': ['sendDataToCloud', 'lasso2d', 'autoScale2d', 'select2d', 'zoom2d', 'pan2d',
                                        'zoomIn2d', 'zoomOut2d', 'resetScale2d', 'toggleSpikelines', 'hoverCompareCartesian',
                                        'hoverClosestCartesian']})
-
-
-def prepare_data_for_shap_plot(x, baseline, prediction, max_vars, rounding_function, digits):
-
-    variable_count = x.shape[0]
-
-    if variable_count > max_vars:
-        last_row = max_vars - 1
-        new_x = x.iloc[0:(last_row + 1), :].copy()
-        new_x.iloc[last_row, new_x.columns.get_loc('variable')] = "+ all other factors"
-        new_x.iloc[last_row, new_x.columns.get_loc('contribution')] = np.sum(
-            x.iloc[last_row:(variable_count - 1), x.columns.get_loc('contribution')])
-
-        x = new_x
-
-    # use for text label and tooltip
-    x.loc[:, 'contribution'] = rounding_function(x.loc[:, 'contribution'], digits)
-    baseline = rounding_function(baseline, digits)
-    prediction = rounding_function(prediction, digits)
-
-    tt = x.apply(lambda row: tooltip_text(row, baseline, prediction), axis=1)
-    x = x.assign(tooltip_text=tt.values)
-
-    lt = label_text(x.iloc[:, x.columns.get_loc("contribution")].tolist())
-    x = x.assign(label_text=lt)
-
-    return x
-
-
-def tooltip_text(row, baseline, prediction):
-    if row.contribution > 0:
-        key_word = "increases"
-    else:
-        key_word = "decreases"
-    return "Average response: " + str(baseline) + "<br>Prediction: " + str(prediction) + "<br>" +\
-           row.variable + "<br>" + key_word + " average response <br>by " + str(np.abs(row.contribution))
-
-
-def label_text(contribution):
-    def to_text(x):
-        if x > 0:
-            return "+" + str(x)
-        else:
-            return str(x)
-
-    return [to_text(c) for c in contribution]
